moved shooter constants and vision constants
