+++ src/main/java/frc/robot/extras/NoteDetector.java
@@ -7,7 +7,7 @@ import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import frc.robot.Constants.VisionConstants;
+import frc.robot.subsystems.vision.VisionConstants;
 import frc.robot.extras.interpolators.BilinearInterpolator;
 
 /** Add your docs here. */
