+++ src/main/java/frc/robot/subsystems/swerve/SwerveDrive.java
@@ -19,12 +19,12 @@ import edu.wpi.first.math.numbers.N3;
 import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import frc.robot.Constants.VisionConstants;
 import frc.robot.extras.Alert;
 import frc.robot.extras.DeviceCANBus;
 import frc.robot.extras.VirtualSubsystem;
 import frc.robot.extras.util.AllianceFlipper;
 import frc.robot.extras.util.TimeUtil;
+import frc.robot.subsystems.vision.VisionConstants;
 import frc.robot.subsystems.swerve.SwerveConstants.DriveConstants;
 import frc.robot.subsystems.swerve.gyroIO.GyroIO;
 import frc.robot.subsystems.swerve.gyroIO.GyroIOInputsAutoLogged;
